add new point evaluation pipeline

+++ tsfc/mixedelement.py
@@ -101,7 +101,7 @@ class MixedElement(object):
                 try:
                     arr = output[d]
                 except KeyError:
-                    arr = numpy.zeros(shape)
+                    arr = numpy.zeros(shape, dtype=tab.dtype)
                     output[d] = arr
 
                 ir = irange[i:i+2]
